requirements.txt
Flask
Flask-JWT-Extended
Flask-RESTful
Flask-SQLAlchemy
paho-mqtt
flask-cors
flask-uploads